requirements.txt
scapy>=2.5.0
python-dotenv>=1.0.0
pymongo>=4.6.0
pika>=1.3.0
requests>=2.31.0
